Revert "gsg"
This reverts commit 8c5b530429e299d469d93ee7668b402f0f940d51.

+++ Server/MirDatabase/QuestProgressInfo.cs
@@ -8,9 +8,16 @@ using Server.MirEnvir;
 
 namespace Server.MirDatabase
 {
+    #region Notes
+    //clean up new flag methods
+    #endregion
+
     public class QuestProgressInfo
     {
-        public int Index;
+        public int Index
+        {
+            get { return Info.Index; }
+        }
 
         public QuestInfo Info;
 
@@ -40,9 +47,9 @@ namespace Server.MirDatabase
 
         public QuestProgressInfo(int index)
         {
-            Index = index;
-
             Info = SMain.Envir.QuestInfoList.FirstOrDefault(e => e.Index == index);
+
+            if (Info == null) return;
 
             foreach (var kill in Info.KillTasks)
                 KillTaskCount.Add(0);
@@ -58,8 +65,9 @@ namespace Server.MirDatabase
 
         public QuestProgressInfo(BinaryReader reader)
         {
-            Index = reader.ReadInt32();
-            Info = SMain.Envir.QuestInfoList.FirstOrDefault(e => e.Index == Index);
+            int index = reader.ReadInt32();
+
+            Info = SMain.Envir.QuestInfoList.FirstOrDefault(e => e.Index == index);
 
             StartDateTime = DateTime.FromBinary(reader.ReadInt64());
             EndDateTime = DateTime.FromBinary(reader.ReadInt64());
@@ -199,8 +207,7 @@ namespace Server.MirDatabase
 
         public bool NeedKill(MonsterInfo mInfo)
         {
-            //if (info.Name != name && !info.Name.Replace(" ", "").StartsWith(name, StringComparison.OrdinalIgnoreCase)) continue;
-            return Info.KillTasks.Where((task, i) => KillTaskCount[i] < task.Count && mInfo.Name.StartsWith(task.Monster.Name, StringComparison.OrdinalIgnoreCase)).Any();
+            return Info.KillTasks.Where((task, i) => KillTaskCount[i] < task.Count && task.Monster == mInfo).Any();
         }
 
         public bool NeedFlag(int flagNumber)
@@ -212,14 +219,14 @@ namespace Server.MirDatabase
 
         #region Process Quest Task
 
-        public void ProcessKill(MonsterInfo mInfo)
+        public void ProcessKill(int mobIndex)
         {
             if (Info.KillTasks.Count < 1) return;
 
             for (int i = 0; i < Info.KillTasks.Count; i++)
             {
-                //if (Info.KillTasks[i].Monster.Index != mobIndex) continue;
-                if (!mInfo.Name.StartsWith(Info.KillTasks[i].Monster.Name, StringComparison.OrdinalIgnoreCase)) continue;
+                if (Info.KillTasks[i].Monster.Index != mobIndex) continue;
+
                 KillTaskCount[i]++;
 
                 return;
@@ -347,7 +354,7 @@ namespace Server.MirDatabase
 
         public void UpdateGotoTask()
         {
-            if (Info.GotoMessage.Length <= 0 || !Completed) return;
+            if (Info.GotoMessage.Length <= 0 || TaskList.Count > 0) return;
 
             TaskList.Add(Info.GotoMessage);
         }
